# backend/docs/swagger.yaml
basePath: /
definitions:
  anydoc.Child:
    properties:
      children:
        items:
          $ref: '#/definitions/anydoc.Child'
        type: array
      value:
        $ref: '#/definitions/anydoc.Value'
    type: object
  anydoc.Value:
    properties:
      file:
        type: boolean
      file_type:
        type: string
      id:
        type: string
      summary:
        type: string
      title:
        type: string
    type: object
...
  consts.CrawlerSource:
    enum:
    - url
    - rss
    - sitemap
    - notion
    - feishu
    - file
    - epub
    - yuque
    - siyuan
    - mindoc
    - wikijs
    - confluence
    type: string
    x-enum-varnames:
    - CrawlerSourceUrl
    - CrawlerSourceRSS
    - CrawlerSourceSitemap
    - CrawlerSourceNotion
    - CrawlerSourceFeishu
    - CrawlerSourceFile
    - CrawlerSourceEpub
    - CrawlerSourceYuque
    - CrawlerSourceSiyuan
    - CrawlerSourceMindoc
    - CrawlerSourceWikijs
    - CrawlerSourceConfluence
  consts.CrawlerStatus:
    enum:
    - pending
    - in_process
    - completed
    - failed
    type: string
    x-enum-varnames:
    - CrawlerStatusPending
    - CrawlerStatusInProcess
    - CrawlerStatusCompleted
    - CrawlerStatusFailed
...
  domain.CreateModelReq:
    properties:
      api_header:
        type: string
      api_key:
        type: string
      api_version:
        description: for azure openai
        type: string
      base_url:
        type: string
      model:
        type: string
      parameters:
        $ref: '#/definitions/github_com_chaitin_panda-wiki_domain.ModelParam'
      provider:
        $ref: '#/definitions/github_com_chaitin_panda-wiki_domain.ModelProvider'
      type:
        allOf:
        - $ref: '#/definitions/domain.ModelType'
        enum:
        - chat
        - embedding
        - rerank
        - analysis
        - analysis-vl
    required:
    - base_url
    - model
    - provider
    - type
    type: object
...
  domain.GetProviderModelListReq:
    properties:
      api_header:
        type: string
      api_key:
        type: string
      base_url:
        type: string
      provider:
        type: string
      type:
        allOf:
        - $ref: '#/definitions/domain.ModelType'
        enum:
        - chat
        - embedding
        - rerank
        - analysis
        - analysis-vl
    required:
    - base_url
    - provider
    - type
    type: object
  domain.GetProviderModelListResp:
    properties:
      models:
        items:
          $ref: '#/definitions/domain.ProviderModelListItem'
        type: array
    type: object
...
  domain.ModelType:
    enum:
    - chat
    - embedding
    - rerank
    - analysis
    - analysis-vl
    type: string
    x-enum-varnames:
    - ModelTypeChat
    - ModelTypeEmbedding
    - ModelTypeRerank
    - ModelTypeAnalysis
    - ModelTypeAnalysisVL
...
  domain.PWResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  domain.ProviderModelListItem:
    properties:
      model:
        type: string
    type: object
...
  domain.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  domain.SwitchModeReq:
    properties:
      auto_mode_api_key:
        description: 百智云 API Key
        type: string
      chat_model:
        description: 自定义对话模型名称
        type: string
      mode:
        enum:
        - manual
        - auto
        type: string
    required:
    - mode
    type: object
  domain.SwitchModeResp:
    properties:
      message:
        type: string
    type: object
...
  domain.UpdateModelReq:
    properties:
      api_header:
        type: string
      api_key:
        type: string
      api_version:
        description: for azure openai
        type: string
      base_url:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      model:
        type: string
      parameters:
        $ref: '#/definitions/github_com_chaitin_panda-wiki_domain.ModelParam'
      provider:
        $ref: '#/definitions/github_com_chaitin_panda-wiki_domain.ModelProvider'
      type:
        allOf:
        - $ref: '#/definitions/domain.ModelType'
        enum:
        - chat
        - embedding
        - rerank
        - analysis
        - analysis-vl
    required:
    - base_url
    - id
    - model
    - provider
    - type
    type: object
...
  github_com_chaitin_panda-wiki_domain.CheckModelReq:
    properties:
      api_header:
        type: string
      api_key:
        type: string
      api_version:
        description: for azure openai
        type: string
      base_url:
        type: string
      model:
        type: string
      parameters:
        $ref: '#/definitions/github_com_chaitin_panda-wiki_domain.ModelParam'
      provider:
        $ref: '#/definitions/github_com_chaitin_panda-wiki_domain.ModelProvider'
      type:
        allOf:
        - $ref: '#/definitions/domain.ModelType'
        enum:
        - chat
        - embedding
        - rerank
        - analysis
        - analysis-vl
    required:
    - base_url
    - model
    - provider
    - type
    type: object
  github_com_chaitin_panda-wiki_domain.CheckModelResp:
    properties:
      content:
        type: string
      error:
        type: string
    type: object
...
  github_com_chaitin_panda-wiki_domain.ModelParam:
    properties:
      context_window:
        type: integer
      max_tokens:
        type: integer
      r1_enabled:
        type: boolean
      support_computer_use:
        type: boolean
      support_images:
        type: boolean
      support_prompt_cache:
        type: boolean
      temperature:
        type: number
    type: object
  github_com_chaitin_panda-wiki_domain.ModelProvider:
    enum:
    - BaiZhiCloud
    type: string
    x-enum-varnames:
    - ModelProviderBrandBaiZhiCloud
...
  v1.CrawlerExportReq:
    properties:
      doc_id:
        type: string
      file_type:
        type: string
      id:
        type: string
      kb_id:
        type: string
      space_id:
        type: string
    required:
    - doc_id
    - id
    - kb_id
    type: object
  v1.CrawlerExportResp:
    properties:
      task_id:
        type: string
    type: object
  v1.CrawlerParseReq:
    properties:
      crawler_source:
        $ref: '#/definitions/consts.CrawlerSource'
      feishu_setting:
        $ref: '#/definitions/v1.FeishuSetting'
      filename:
        type: string
      kb_id:
        type: string
      key:
        type: string
    required:
    - crawler_source
    - kb_id
    type: object
  v1.CrawlerParseResp:
    properties:
      docs:
        $ref: '#/definitions/anydoc.Child'
      id:
        type: string
    type: object
  v1.CrawlerResultItem:
    properties:
      content:
        type: string
      status:
        $ref: '#/definitions/consts.CrawlerStatus'
      task_id:
        type: string
    type: object
  v1.CrawlerResultReq:
    properties:
      task_id:
        type: string
    required:
    - task_id
    type: object
  v1.CrawlerResultResp:
    properties:
      content:
        type: string
      status:
        $ref: '#/definitions/consts.CrawlerStatus'
    required:
    - status
    type: object
  v1.CrawlerResultsReq:
    properties:
      task_ids:
        items:
          type: string
        type: array
    required:
    - task_ids
    type: object
  v1.CrawlerResultsResp:
    properties:
      list:
        items:
          $ref: '#/definitions/v1.CrawlerResultItem'
        type: array
      status:
        $ref: '#/definitions/consts.CrawlerStatus'
    type: object
...
  v1.FeishuSetting:
    properties:
      app_id:
        type: string
      app_secret:
        type: string
      space_id:
        type: string
      user_access_token:
        type: string
    type: object
...
info:
  contact: {}
  description: panda-wiki API documentation
  title: panda-wiki API
  version: "1.0"
paths:
...
  /api/v1/crawler/export:
    post:
      consumes:
      - application/json
      description: CrawlerExport
      parameters:
      - description: Scrape
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.CrawlerExportReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.PWResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.CrawlerExportResp'
              type: object
      summary: CrawlerExport
      tags:
      - crawler
  /api/v1/crawler/parse:
    post:
      consumes:
      - application/json
      description: 解析文档树
      parameters:
      - description: Scrape
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.CrawlerParseReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.PWResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.CrawlerParseResp'
              type: object
      summary: 解析文档树
      tags:
      - crawler
  /api/v1/crawler/result:
    get:
      consumes:
      - application/json
      description: Retrieve the result of a previously started scraping task
      parameters:
      - description: Crawler Result Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.CrawlerResultReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.PWResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.CrawlerResultResp'
              type: object
      summary: Get Crawler Result
      tags:
      - crawler
  /api/v1/crawler/results:
    post:
      consumes:
      - application/json
      description: Retrieve the results of a previously started scraping task
      parameters:
      - description: Crawler Results Request
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/v1.CrawlerResultsReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.PWResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.CrawlerResultsResp'
              type: object
      summary: Get Crawler Results
      tags:
      - crawler
...
  /api/v1/model:
    post:
      consumes:
      - application/json
      description: create model
      parameters:
      - description: create model request
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/domain.CreateModelReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: create model
      tags:
      - model
    put:
      consumes:
      - application/json
      description: update model
      parameters:
      - description: update model request
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateModelReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      tags:
      - model
  /api/v1/model/check:
    post:
      consumes:
      - application/json
      description: check model
      parameters:
      - description: check model request
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/github_com_chaitin_panda-wiki_domain.CheckModelReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_chaitin_panda-wiki_domain.CheckModelResp'
              type: object
      summary: check model
      tags:
      - model
...
  /api/v1/model/provider/supported:
    post:
      consumes:
      - application/json
      description: get provider supported model list
      parameters:
      - description: get supported model list request
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/domain.GetProviderModelListReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.PWResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.GetProviderModelListResp'
              type: object
      summary: get provider supported model list
      tags:
      - model
  /api/v1/model/switch-mode:
    post:
      consumes:
      - application/json
      description: switch model mode between manual and auto
      parameters:
      - description: switch mode request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.SwitchModeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.SwitchModeResp'
              type: object
      summary: switch mode
      tags:
      - model
...
securityDefinitions:
  bearerAuth:
    description: Type "Bearer" + a space + your token to authorize
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"